# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BuntesBegegnen.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/admin/users/{userId}/roles": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/users/{userId}/locked": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isLocked",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/users/{userId}/reset-password": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/admin/users/{userId}": {
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Activity": {
        "required": [
          "category",
          "createdAt",
          "createdById",
          "description",
          "endTime",
          "isAllDay",
          "location",
          "maxNumberOfParticipants",
          "promoterId",
          "recurrenceFrequency",
          "recurrenceInterval",
          "registrationLocked",
          "repeatCount",
          "repeatUntil",
          "startTime",
          "title",
          "visibility"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/User"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "promoterId": {
            "type": "integer",
            "format": "int32"
          },
          "promoter": {
            "$ref": "#/components/schemas/Promoter"
          },
          "title": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/ActivityVisibility"
          },
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "isAllDay": {
            "type": "boolean"
          },
          "maxNumberOfParticipants": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "registrationLocked": {
            "type": "boolean"
          },
          "location": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "recurrenceFrequency": {
            "$ref": "#/components/schemas/ActivityRecurrenceFrequency"
          },
          "recurrenceInterval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "repeatUntil": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "repeatCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "recurrenceDates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceDate"
            }
          },
          "recurrenceByDay": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceByDay"
            }
          },
          "recurrenceByMonthDay": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceByMonthDay"
            }
          },
          "recurrenceExceptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityRecurrenceException"
            }
          }
        },
        "additionalProperties": false
      },
      "ActivityCategory": {
        "enum": [
          "Excursion",
          "Creativity",
          "Exercise"
        ],
        "type": "string"
      },
...
      "ActivityRecurrenceByDay": {
        "required": [
          "activityId",
          "dayOfWeek",
          "ordinal"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "ordinal": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          }
        },
        "additionalProperties": false
      },
...
      "ActivityRecurrenceByMonthDay": {
        "required": [
          "activityId",
          "monthDay"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "monthDay": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ActivityRecurrenceDate": {
        "required": [
          "activityId",
          "startTime"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ActivityRecurrenceException": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ActivityRecurrenceFrequency": {
        "enum": [
          "None",
          "Weekly",
          "Monthly",
          "FixedDates"
        ],
        "type": "string"
      },
      "ActivityVisibility": {
        "enum": [
          "PrivateDraft",
          "SharedDraft",
          "Public"
        ],
        "type": "string"
      },
...
      "DayOfWeek": {
        "enum": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday"
        ],
        "type": "string"
      },
...
      "Promoter": {
        "required": [
          "city",
          "createdAt",
          "createdById",
          "houseNumber",
          "name",
          "streetName",
          "website",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "$ref": "#/components/schemas/User"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "streetName": {
            "type": "string",
            "nullable": true
          },
          "houseNumber": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "required": [
          "createdAt",
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDay": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "streetName": {
            "type": "string",
            "nullable": true
          },
          "houseNumber": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "promoterId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "promoter": {
            "$ref": "#/components/schemas/Promoter"
          },
          "goals": {
            "type": "string",
            "nullable": true
          },
          "hobbies": {
            "type": "string",
            "nullable": true
          },
          "favoriteCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserFavoriteCategory"
            }
          },
          "impairedSight": {
            "type": "boolean"
          },
          "impairedHearing": {
            "type": "boolean"
          },
          "impairedSpeech": {
            "type": "boolean"
          },
          "impairedMobility": {
            "type": "boolean"
          },
          "additionalHandicaps": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserFavoriteCategory": {
        "required": [
          "category",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "category": {
            "$ref": "#/components/schemas/ActivityCategory"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter the authentication token (without the 'Bearer' prefix)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}